+++ examples/ecs/dotnet/MyStack.cs
@@ -13,6 +13,7 @@ class MyStack : Stack
         var service = new Awsx.Ecs.FargateService("my-service", new Awsx.Ecs.FargateServiceArgs
         {
             Cluster = cluster.Arn,
+            AssignPublicIp = true,
             TaskDefinitionArgs = new Awsx.Ecs.Inputs.FargateServiceTaskDefinitionArgs
             {
                 Container = new Awsx.Ecs.Inputs.TaskDefinitionContainerDefinitionArgs